+++ NEWS.md
@@ -4,6 +4,7 @@
 * Allow `esquisse_server()` to be initialized with NULL without forcing the import module to appear using `import_from = NULL` [#232](https://github.com/dreamRs/esquisse/issues/232).
 * New geom available in the interface : `geom_path`.
 * `esquisse_ui()`: ability to display not element of the module (header's button, play/pause button, export options).
+* `safe_ggplot()` has a new argument `show_notification` to allow not displaying notifications or displaying theme only once per session.
 
 * i18n: new translations added:
  + polish, activate with `set_i18n("pl")`, thanks to [@jakub-jedrusiak](https://github.com/jakub-jedrusiak)